+++ src/main/java/net/okocraft/spawners/Configuration.java
@@ -1,4 +1,4 @@
-package com.github.siroshun09.spawners;
+package net.okocraft.spawners;
 
 import com.github.siroshun09.sirolibrary.config.BukkitConfig;
 import com.github.siroshun09.sirolibrary.message.BukkitMessage;
