+++ src/main/java/net/okocraft/spawners/Permissions.java
@@ -1,4 +1,4 @@
-package com.github.siroshun09.spawners;
+package net.okocraft.spawners;
 
 import org.bukkit.Bukkit;
 import org.bukkit.permissions.Permission;
