+++ src/main/java/net/okocraft/spawners/Spawners.java
@@ -1,10 +1,10 @@
-package com.github.siroshun09.spawners;
+package net.okocraft.spawners;
 
 import com.github.siroshun09.sirolibrary.bukkitutils.BukkitUtil;
-import com.github.siroshun09.spawners.listeners.CommandListener;
-import com.github.siroshun09.spawners.listeners.EntityListener;
-import com.github.siroshun09.spawners.listeners.SpawnerListener;
-import com.github.siroshun09.spawners.stack.MobStacker;
+import net.okocraft.spawners.listeners.CommandListener;
+import net.okocraft.spawners.listeners.EntityListener;
+import net.okocraft.spawners.listeners.SpawnerListener;
+import net.okocraft.spawners.stack.MobStacker;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
