+++ src/main/java/net/okocraft/spawners/events/MobStackEvent.java
@@ -1,4 +1,4 @@
-package com.github.siroshun09.spawners.events;
+package net.okocraft.spawners.events;
 
 import org.bukkit.entity.Mob;
 import org.bukkit.event.Cancellable;
@@ -6,18 +6,20 @@ import org.bukkit.event.HandlerList;
 import org.bukkit.event.entity.EntityEvent;
 import org.jetbrains.annotations.NotNull;
 
-public class MobUnstackEvent extends EntityEvent implements Cancellable {
+public class MobStackEvent extends EntityEvent implements Cancellable {
     private static final HandlerList handlerList = new HandlerList();
 
     private final Mob target;
+    private final Mob stacked;
     private int amount;
     private boolean isCancelled;
 
 
-    public MobUnstackEvent(@NotNull Mob target, int amount) {
+    public MobStackEvent(@NotNull Mob target, @NotNull Mob stacked, int amount) {
         super(target);
         this.target = target;
         this.amount = amount;
+        this.stacked = stacked;
         isCancelled = false;
     }
 
@@ -28,6 +30,10 @@ public class MobUnstackEvent extends EntityEvent implements Cancellable {
 
     public Mob getTarget() {
         return target;
+    }
+
+    public Mob getStacked() {
+        return stacked;
     }
 
     public int getAmount() {
