+++ src/main/java/net/okocraft/spawners/listeners/SpawnerListener.java
@@ -1,11 +1,11 @@
-package com.github.siroshun09.spawners.listeners;
+package net.okocraft.spawners.listeners;
 
 import com.github.siroshun09.sirolibrary.bukkitutils.BukkitUtil;
-import com.github.siroshun09.spawners.Configuration;
-import com.github.siroshun09.spawners.Messages;
-import com.github.siroshun09.spawners.Spawners;
-import com.github.siroshun09.spawners.events.SpawnerPlaceEvent;
-import com.github.siroshun09.spawners.stack.MobStacker;
+import net.okocraft.spawners.Configuration;
+import net.okocraft.spawners.Messages;
+import net.okocraft.spawners.Spawners;
+import net.okocraft.spawners.events.SpawnerPlaceEvent;
+import net.okocraft.spawners.stack.MobStacker;
 import org.bukkit.ChatColor;
 import org.bukkit.GameMode;
 import org.bukkit.Material;
