+++ src/main/java/net/okocraft/spawners/stack/StackKill.java
@@ -1,7 +1,7 @@
-package com.github.siroshun09.spawners.stack;
+package net.okocraft.spawners.stack;
 
 import com.github.siroshun09.sirolibrary.message.BukkitMessage;
-import com.github.siroshun09.spawners.Messages;
+import net.okocraft.spawners.Messages;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
