+++ src/main/java/net/okocraft/spawners/stack/StackTask.java
@@ -1,6 +1,6 @@
-package com.github.siroshun09.spawners.stack;
+package net.okocraft.spawners.stack;
 
-import com.github.siroshun09.spawners.Configuration;
+import net.okocraft.spawners.Configuration;
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.entity.Ageable;
